refactor: Refactored event DTO's and models
It was a mess, now it's slightly better
