+++ Singer.API/ClientApp/package.json
@@ -1,6 +1,6 @@
 {
    "name": "Singer",
-   "version": "0.1.0-issue-AB-59-registration.1012",
+   "version": "0.1.0-issue-registrations-overview.1015",
    "scripts": {
       "ng": "ng",
       "start": "ng serve",
