+++ Singer.API/ClientApp/src/app/modules/admin/components/singerevents/singer-event-registrations/singer-event-registrations.component.ts
@@ -1,5 +1,8 @@
-import { Component, OnInit, Inject, Output, EventEmitter } from '@angular/core';
-import { SingerEvent } from 'src/app/modules/core/models/singerevent.model';
+import { Component, OnInit, Inject } from '@angular/core';
+import {
+   SingerEvent,
+   SingerEventLocation,
+} from 'src/app/modules/core/models/singerevent.model';
 import { FormGroup } from '@angular/forms';
 import {
    MatDialogRef,
@@ -9,14 +12,12 @@ import {
    MatSelectChange,
 } from '@angular/material';
 import { SingerEventsService } from 'src/app/modules/core/services/singerevents-api/singerevents.service';
-import { Registrant } from 'src/app/modules/core/models/registrant.model';
 import { EventSlot } from 'src/app/modules/core/models/eventslot';
 import { RegistrationStatus } from 'src/app/modules/core/models/enum';
 
 import { SingerAdminEventService } from '../../../services/singer-admin-event.service';
 import { SingerEventLocationService } from 'src/app/modules/core/services/singerevents-api/singerevent-location.service';
-import { SingerEventLocation } from 'src/app/modules/core/models/singer-event-location.dto';
-import { DaycareLocationDTO } from 'src/app/modules/core/models/daycarelocation.dto';
+import { DaycareLocationDTO } from 'src/app/modules/core/DTOs/daycarelocation.dto';
 
 export class SingerEventRegistrationData {
    event: SingerEvent;
@@ -45,7 +46,8 @@ export class SingerEventRegistrationsComponent implements OnInit {
       @Inject(MAT_DIALOG_DATA) data: SingerEventRegistrationData
    ) {
       this.event = data.event;
-      this.hasDaycare =data.event.hasDayCareAfter || data.event.hasDayCareBefore;
+      this.hasDaycare =
+         data.event.hasDayCareAfter || data.event.hasDayCareBefore;
       this.formGroup = new FormGroup({});
    }
 
@@ -66,7 +68,8 @@ export class SingerEventRegistrationsComponent implements OnInit {
                      r.id,
                      r.startDateTime,
                      r.endDateTime,
-                     r.registrations
+                     r.registrations,
+                     r.registrations.length
                   )
             );
 
@@ -85,18 +88,18 @@ export class SingerEventRegistrationsComponent implements OnInit {
                   : this.eventSlots[0];
          });
 
-         this._singerEventLocationService.fetchSingerEventLocationsData('asc', 'name', 0, 1000, '')
+      this._singerEventLocationService
+         .fetchSingerEventLocationsData('asc', 'name', 0, 1000, '')
          .subscribe(res => {
             this.availableLocations = res.items as SingerEventLocation[];
          });
    }
 
-
    compareLocations(
       locationX: SingerEventLocation,
       locationY: DaycareLocationDTO
    ) {
-      if(!locationY){
+      if (!locationY) {
          return false;
       }
       return locationX.id === locationY.id;
@@ -110,7 +113,9 @@ export class SingerEventRegistrationsComponent implements OnInit {
             registrationId,
             daycareLocation.id
          )
-         .subscribe(res => this._snackBar.open(`Opvanglocatie naar ${res.name} gewijzigd`));
+         .subscribe(res =>
+            this._snackBar.open(`Opvanglocatie naar ${res.name} gewijzigd`)
+         );
    }
 
    changeRegistration(event: MatButtonToggleChange, registrationId: string) {
