+++ Singer.API/ClientApp/src/app/modules/admin/services/singer-admin-event.proxy.ts
@@ -3,7 +3,7 @@ import { ApiService } from '../../core/services/api.service';
 import { map } from 'rxjs/operators';
 import { Observable } from 'rxjs';
 import { RegistrationStatus } from '../../core/models/enum';
-import { DaycareLocationDTO } from '../../core/models/daycarelocation.dto';
+import { DaycareLocationDTO } from '../../core/DTOs/daycarelocation.dto';
 
 @Injectable({
    providedIn: 'root',
