+++ Singer.API/ClientApp/src/app/modules/core/DTOs/pagination.dto.ts
@@ -1,8 +1,8 @@
-export interface PaginationDTO {
+export interface PaginationDTO<TDTO = any> {
    previousPageUrl: string;
    nextPageUrl: string;
    currentPageUrl: string;
    size: number;
    totalSize: number;
-   items: any[];
+   items: TDTO[];
 }
