+++ Singer.API/ClientApp/src/app/modules/core/DTOs/singer-event-location.dto.ts
@@ -1,12 +1,3 @@
-export class SingerEventLocation {
-   id: string;
-   name: string;
-   address: string;
-   postalCode: string;
-   city: string;
-   country: string;
-}
-
 export interface SingerEventLocationDTO {
    id: string;
    name: string;
