+++ Singer.API/ClientApp/src/app/modules/core/components/single-registration/single-registration.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { EventSlotRegistrations, RegistrationStatus } from '../../models/singerevent.model';
+import { EventSlotRegistrations } from '../../models/singerevent.model';
 import { Registrant } from '../../models/registrant.model';
 import { SingerEventsService } from '../../services/singerevents-api/singerevents.service';
 import { MatSnackBar } from '@angular/material';
+import { RegistrationStatus } from '../../models/enum';
 
 @Component({
    selector: 'app-single-registration',
@@ -13,32 +14,29 @@ export class SingleRegistrationComponent implements OnInit {
    @Input() eventId: string;
    @Input() eventSlots: EventSlotRegistrations[] = [];
    @Input() careUsers: Registrant[] = [];
-   constructor(private _eventService: SingerEventsService, private _snackBar: MatSnackBar) {}
+   constructor(
+      private _eventService: SingerEventsService,
+      private _snackBar: MatSnackBar
+   ) {}
 
    ngOnInit() {}
 
    getEventRegistrationStatus(careUserId: string): RegistrationStatus {
-      return this.careUsers.find(x => x.careUserId === careUserId).registrationStatus;
+      return this.careUsers.find(x => x.careUserId === careUserId)
+         .registrationStatus;
    }
 
-   registerCareUserOnEvent(
-      careUser: Registrant
-   ) {
+   registerCareUserOnEvent(careUser: Registrant) {
       this._eventService
-         .registerCareUserOnEvent(
-            this.eventId,
-            careUser.careUserId
-         )
+         .registerCareUserOnEvent(this.eventId, careUser.careUserId)
          .subscribe(
             res => {
-               this._snackBar.open(
-                  `${careUser.name} werd ingeschreven`,
-                  'OK',
-                  { duration: 2000 }
-               );
-               this.careUsers
-                  .find(x => x.careUserId === careUser.careUserId).registrationStatus = res[0].status;
-
+               this._snackBar.open(`${careUser.name} werd ingeschreven`, 'OK', {
+                  duration: 2000,
+               });
+               this.careUsers.find(
+                  x => x.careUserId === careUser.careUserId
+               ).registrationStatus = res[0].status;
             },
             err => {
                console.log(err);
@@ -47,5 +45,3 @@ export class SingleRegistrationComponent implements OnInit {
          );
    }
 }
-
-
