+++ Singer.API/ClientApp/src/app/modules/core/models/careuser.model.ts
@@ -3,7 +3,7 @@ import {
    LinkedLegalGuardianDTO,
    LinkedLegalGuardian,
 } from './legalguardian.model';
-import { SingerEventLocation } from './singer-event-location.dto';
+import { SingerEventLocation } from './singerevent.model';
 
 export interface CareUserDTO {
    id: string;
