+++ Singer.API/ClientApp/src/app/modules/core/models/daycarelocation.model.ts
@@ -1,0 +1,4 @@
+export class DaycareLocation {
+   id: string;
+   name: string;
+}
