+++ Singer.API/ClientApp/src/app/modules/core/models/eventslot.dto.ts
@@ -1,5 +1,4 @@
 export interface EventSlotDTO {
-   id: string;
    startDateTime: Date;
    endDateTime: Date;
    currentRegistrants: number;
