+++ Singer.API/ClientApp/src/app/modules/core/services/care-users-api/careusers.service.ts
@@ -7,7 +7,7 @@ import {
    CareUser,
    CreateCareUserDTO,
 } from '../../models/careuser.model';
-import { PaginationDTO } from '../../models/pagination.model';
+import { PaginationDTO } from '../../DTOs/pagination.dto';
 @Injectable({
    providedIn: 'root',
 })
