+++ Singer.API/ClientApp/src/app/modules/core/services/legal-guardians-api/legalguardians.service.ts
@@ -7,7 +7,7 @@ import {
    CreateLegalGuardianDTO,
    LegalGuardian,
 } from '../../models/legalguardian.model';
-import { PaginationDTO } from '../../models/pagination.model';
+import { PaginationDTO } from '../../DTOs/pagination.dto';
 
 @Injectable({
    providedIn: 'root',
