+++ Singer.API/ClientApp/src/app/modules/dashboard/components/event-search/event-search.component.ts
@@ -1,8 +1,8 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { FormGroup, FormControl } from '@angular/forms';
 import { MAT_DATE_FORMATS } from '@angular/material';
-import { SingerEventLocation } from 'src/app/modules/core/models/singer-event-location.dto';
 import { MY_FORMATS } from 'src/app/modules/core/core.module';
+import { SingerEventLocation } from 'src/app/modules/core/models/singerevent.model';
 
 @Component({
    selector: 'app-event-search',
